Move stuff around a bit
