+++ packages/@headlessui-react/src/utils/machine.ts
